# gateway/internal/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "TITAN Gateway API",
    "description": "Zero-Trust LLM Security Gateway. The Admin API (`/admin/v1/*`) manages tenants, API keys, policies and audit logs and is gated by the `X-Admin-Token` header. The data-plane proxy (`/v1/*`) forwards OpenAI-compatible requests to the upstream LLM after the seven-stage inspection pipeline, and is gated by a tenant API key. The read API (`/api/*`) exposes non-sensitive metrics.",
    "version": "1.0.0",
    "license": { "name": "MIT" }
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "Local development" }
  ],
...
  "paths": {
...
    "/api/events": {
      "get": {
        "tags": ["Metrics"],
        "summary": "Recent threat / decision events",
        "security": [],
        "parameters": [ { "name": "n", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 }, "description": "Number of events to return" } ],
        "responses": { "200": { "description": "Cluster-wide event feed", "content": { "application/json": { "schema": { "type": "object", "properties": { "events": { "type": "array", "items": { "type": "object" } }, "count": { "type": "integer" } } } } } } }
      }
    },
...
    "/v1/batch": {
      "post": {
        "tags": ["Batch"],
        "summary": "Submit an asynchronous batch of prompts",
        "description": "Each request in the batch is run through the ML governance gate (ALLOW / MASK / BLOCK) and then forwarded to the upstream LLM. Returns immediately with a job id; poll the status URL for results. Max 100 items per batch.",
        "security": [ { "ApiKeyAuth": [] } ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["requests"], "properties": { "requests": { "type": "array", "maxItems": 100, "items": { "$ref": "#/components/schemas/ChatCompletionRequest" } } } } } } },
        "responses": {
          "202": { "description": "Batch accepted", "content": { "application/json": { "schema": { "type": "object", "properties": { "job_id": { "type": "string", "format": "uuid" }, "status": { "type": "string" }, "total": { "type": "integer" }, "status_url": { "type": "string" } } } } } },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
    "/v1/batch/{id}": {
      "get": {
        "tags": ["Batch"],
        "summary": "Get batch job status and results",
        "description": "Scoped to the calling tenant; another tenant's job id returns 404.",
        "security": [ { "ApiKeyAuth": [] } ],
        "parameters": [ { "name": "id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } } ],
        "responses": {
          "200": { "description": "Batch job", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BatchJob" } } } },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AdminToken": { "type": "apiKey", "in": "header", "name": "X-Admin-Token", "description": "Master admin secret for /admin/v1/* routes." },
      "ApiKeyAuth": { "type": "http", "scheme": "bearer", "description": "Tenant API key for the /v1/* data plane." }
    },
    "responses": {
      "Unauthorized": { "description": "Missing or invalid credential", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "BadRequest": { "description": "Invalid request body or parameter", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "NotFound": { "description": "Resource not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    },
    "schemas": {
      "Error": { "type": "object", "properties": { "error": { "type": "string" } } },
...
      "BatchJob": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "tenant_id": { "type": "string", "format": "uuid" },
          "status": { "type": "string", "enum": ["queued", "processing", "completed"] },
          "total": { "type": "integer" },
          "completed": { "type": "integer" },
          "created_at": { "type": "string", "format": "date-time" },
          "completed_at": { "type": "string", "format": "date-time", "nullable": true },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "index": { "type": "integer" },
                "status": { "type": "string", "enum": ["queued", "allowed", "masked", "blocked", "error"] },
                "risk_score": { "type": "number" },
                "reason": { "type": "string" },
                "response": { "type": "object", "description": "Upstream LLM response (present when allowed/masked)" },
                "error": { "type": "string" }
              }
            }
          }
        }
      },
      "ChatCompletionRequest": {
        "type": "object",
        "required": ["model", "messages"],
        "properties": {
          "model": { "type": "string", "example": "llama-3.3-70b-versatile" },
          "messages": {
            "type": "array",
            "items": { "type": "object", "properties": { "role": { "type": "string", "enum": ["system", "user", "assistant"] }, "content": { "type": "string" } } }
          },
          "temperature": { "type": "number" },
          "stream": { "type": "boolean" }
        }
      }
    }
  },
  "security": [ { "AdminToken": [] } ]
}